Changed object database profoundly so that pointers are never exposed to the user.  This way, all operations are guaranteed to complete with data integrity.

+++ utils_common.h
@@ -191,7 +191,7 @@ no_space (int rv)
 typedef union datum_u {
 
     void *pointer;
-    int integer;
+    int64 integer;
 
 } datum_t;
